feat: gridSkeleton로 따로 분리
